added small optimization for moving pieces
